+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/DefaultComponents.kt
@@ -60,6 +60,7 @@ import mozilla.components.feature.readerview.ReaderViewMiddleware
 import mozilla.components.feature.search.SearchUseCases
 import mozilla.components.feature.session.HistoryDelegate
 import mozilla.components.feature.session.SessionUseCases
+import mozilla.components.feature.sitepermissions.SitePermissionsStorage
 import mozilla.components.feature.tabs.TabsUseCases
 import mozilla.components.feature.webnotifications.WebNotificationFeature
 import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient
@@ -115,6 +116,8 @@ open class DefaultComponents(private val applicationContext: Context) {
 
     private val sessionStorage by lazy { SessionStorage(applicationContext, engine) }
 
+    private val permissionStorage by lazy { SitePermissionsStorage(applicationContext) }
+
     val thumbnailStorage by lazy { ThumbnailStorage(applicationContext) }
 
     val store by lazy {
@@ -147,7 +150,7 @@ open class DefaultComponents(private val applicationContext: Context) {
                 .enable()
 
             WebNotificationFeature(applicationContext, engine, icons, R.drawable.ic_notification,
-                BrowserActivity::class.java)
+                permissionStorage, BrowserActivity::class.java)
         }
     }